Refactor tensor handling in metrics and unlearning modules
- Update tensor conversion to use .float() before numpy conversion in utils.py, min_k_plus_plus.py, and min_k.py for improved numerical stability.
- Modify compute_loss method in grad_ascent.py to include an additional parameter for batch size.

+++ src/evals/metrics/mia/min_k_plus_plus.py
@@ -30,9 +30,9 @@ class MinKPlusPlusAttack(MinKProbAttack):
 
         # Handle numerical stability
         sigma = torch.clamp(sigma, min=1e-6)
-        scores = (target_prob.cpu().numpy() - mu.cpu().numpy()) / torch.sqrt(
+        scores = (target_prob.cpu().float().numpy() - mu.cpu().float().numpy()) / torch.sqrt(
             sigma
-        ).cpu().numpy()
+        ).cpu().float().numpy()
 
         # Take bottom k% as the attack score
         num_k = max(1, int(len(scores) * self.k))
